+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Fixed
 - Correct ProjectItem's created date label display (@jswk)
+- Proper displaying questions in the resource opinion form (@goreck888)
 
 ### Security